# ws-server/api/openapi.yaml
openapi: 3.0.0
info:
  title: TaNoClo API Documentation
  description: >
    Server documentation for TaNoClo. Fully compatible with community
    integrations like pytado and Home Assistant.

    This specification builds upon the community-managed Tado OpenAPI v2
    specification by kritsel (available at
    https://github.com/kritsel/tado-openapispec-v2).
  version: 2026.09.11.1
servers:
  - url: /api/v2
    description: TaNoClo API v2 base
  - url: /
    description: TaNoClo root base
security:
  - oauth2: []
paths:
...
  /homes/{homeId}/zones/{zoneId}/details:
    put:
      summary: Change the name of the zone
      operationId: setDetails
      tags:
        - zone
      parameters:
        - in: path
          name: homeId
          schema:
            $ref: "#/components/schemas/HomeId"
          required: true
          description: unique ID of a home the authenticated user has access to
        - in: path
          name: zoneId
          schema:
            $ref: "#/components/schemas/ZoneId"
          required: true
          description: ID of a zone (unique within a home)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneDetailsInput"
      responses:
        "200":
          description: successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "401":
          $ref: "#/components/responses/Unauthorized401"
        "403":
          $ref: "#/components/responses/AccessDenied403"
        "404":
          $ref: "#/components/responses/NotFound404"
        "422":
          $ref: "#/components/responses/InputError422"
    get:
      summary: Get details of the zone
      operationId: getZoneDetails
      tags:
        - zone
      parameters:
        - in: path
          name: homeId
          schema:
            $ref: "#/components/schemas/HomeId"
          required: true
          description: unique ID of a home
        - in: path
          name: zoneId
          schema:
            $ref: "#/components/schemas/ZoneId"
          required: true
          description: ID of a zone
      responses:
        "200":
          description: successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "401":
          $ref: "#/components/responses/Unauthorized401"
        "403":
          $ref: "#/components/responses/AccessDenied403"
        "404":
          $ref: "#/components/responses/NotFound404"
...
  /homes/{homeId}/zones/{zoneId}/control:
    get:
      summary: >
        Returns the heating circuit the zone (a.k.a. room) belongs to, and the
        tado devices within the zone  grouped by their duty. 
      description: >
        There are multiple ways to list the tado devices in a home or zone:

        * GET /homes/{homeId}/devices 

        * GET /homes/{homeId}/deviceList 

        * GET /homes/{homeId}/zones

        * GET /homes/{homeId}/zones/{zoneId}/control **this operation**


        Distinguishing features of **this operation**:

        * **device types**: devices of DeviceType 'IB01' (Internet Bridge) and
        'BR02' (Wireless Receiver) are not included in the response

        * **duties**: the response tells you thy duty each device has (ui,
        driver, leader)

        * **zone**: the response tells you the zone each device belongs to
      operationId: getZoneControl
      tags:
        - device
        - heating circuit
      parameters:
        - in: path
          name: homeId
          schema:
            $ref: "#/components/schemas/HomeId"
          required: true
          description: unique ID of a home the authenticated user has access to
        - in: path
          name: zoneId
          schema:
            $ref: "#/components/schemas/ZoneId"
          required: true
          description: ID of a zone (unique within a home)
      responses:
        "200":
          description: successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneControl"
        "401":
          $ref: "#/components/responses/Unauthorized401"
        "403":
          $ref: "#/components/responses/AccessDenied403"
        "404":
          $ref: "#/components/responses/NotFound404"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: /oauth/token
          scopes: {}
  responses:
    Unauthorized401:
      description: unauthorized, invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AccessDenied403:
      description: access denied, the authenticated user does not have access to the
        specified resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound404:
      description: not found, the specified resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InputError422:
      description: >
        The request body is invalid or the requested operation is not supported.
        This can have all kinds of reasons, some examples:

        * a required property is missing

        * an invalid value is used for a property

        * a particular operation is not allowed on a resource given the state of
        the resource

        * a particular zone specific operation is not supported for the zone's
        ZoneTYpe

        * a particular operation requires a paid Auto-Assist subscription which
        is not active for the home
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse422"
  schemas:
...
    BatteryState:
      type: string
      enum:
        - LOW
        - NORMAL
      example: NORMAL
      description: device property known to be present for DeviceType SU02, VA02
...
    Device:
      type: object
      description: >
        A piece of tado hardware


        Some of the properties are only available for certain device
        types.  Certain values for some properties are only applicable to
        certain device types.  Wherever device type specific information is
        known, it is documented.  But you cannot count on this information being
        fully complete.
      properties:
        deviceType:
          $ref: "#/components/schemas/DeviceType"
        serialNo:
          $ref: "#/components/schemas/DeviceId"
        shortSerialNo:
          type: string
          example: SU4126044345
        currentFwVersion:
          type: string
          example: 215.1
        connectionState:
          type: object
          description: property known to be present for DeviceType SU02, VA02, BR02
          properties:
            value:
              type: boolean
              example: true
            timestamp:
              type: string
              format: date-time
              example: 2022-11-25T09:40:14.803Z
        characteristics:
          type: object
          properties:
            capabilities:
              type: array
              description: array known to be empty for DeviceType BR02
              items:
                type: string
                example: INSIDE_TEMPERATURE_MEASUREMENT
                description: >
                  known values: 

                  * INSIDE_TEMPERATURE_MEASUREMENT (seen for DeviceType SU02,
                  VA02)

                  * IDENTIFY (seen for DeviceType SU02, VA02)

                  * RADIO_ENCRYPTION_KEY_ACCESS (seen for DeviceType IB01)
        mountingState:
          type: object
          description: property known to be present for DeviceType VA02
          properties:
            value:
              type: string
              example: CALIBRATED
              description: |
                known values: 
                * CALIBRATED
                * CALIBRATING
            timestamp:
              type: string
              format: date-time
              example: 2022-11-25T09:40:14.803Z
        mountingStateWithError:
          type: string
          example: CALIBRATED
          description: |
            property known to be present for DeviceType VA02; 
            known values: 
            * CALIBRATED
            * CALIBRATING
        batteryState:
          $ref: "#/components/schemas/BatteryState"
        orientation:
          $ref: "#/components/schemas/Orientation"
        childLockEnabled:
          type: boolean
          example: false
          description: property known to be present for DeviceType VA02
        isDriverConfigured:
          type: boolean
          example: true
          description: property known to be present for DeviceType BR02
        inPairingMode:
          type: boolean
          example: false
          description: property known to be present for DeviceType IB01
        accessPointWiFi:
          type: object
          description: property known to be present for DeviceType WR02 (AC control)
          properties:
            ssid:
              type: string
              example: tado1234
        commandTableUploadState:
          type: string
          example: FAILED
          description: property known to be present for DeviceType WR02 (AC control)
    DeviceExtra:
      description: >
        A piece of tado hardware.


        Some of the properties are only available for certain device
        types.  Certain values for some properties are only applicable for
        certain device types.  Wherever device type specific information is
        known, it is documented.
      allOf:
        - $ref: "#/components/schemas/Device"
        - properties:
            duties:
              type: array
              items:
                type: string
                example: ZONE_UI
                description: >
                  known values: 

                  * ZONE_UI (seen for DeviceType SU02, VA02)

                  * ZONE_LEADER (seen for DeviceType SU02, VA02; 
                    indicates that the device is the temperature measuring device for the zone)

                  * ZONE_DRIVER (seen for DeviceType VA02)
    DeviceId:
      type: string
      example: SU4126044356
      description: the serialNo of a tado device
...
    DeviceType:
      type: string
      example: SU02
      description: |
        data element of Device

        known values: 
        * GW01 (Gateway V1)
        * GW02 (Bridge V2)
        * IB01 (Internet Bridge V3+)
        * BX02 (Box V1)
        * BU01 (Extension Kit UK)
        * EK01 (Extension Kit UK)
        * BR02 (Wireless Receiver V3+)
        * BP02 (Wireless Receiver UK V3+)
        * RU01 (Smart Thermostat V3)
        * RU02 (Wired Smart Thermostat V3+)
        * TS02 (Temp Sensor V1)
        * SU02 (Wireless Temperature Sensor V3+)
        * VA01 (Smart Radiator Thermostat V3)
        * VA02 (Smart Radiator Thermostat V3+)
        * WR01 (Smart AC Control V3)
        * WR02 (Smart AC Control V3+)
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: meaningful textual code associated with the HTTP status code like
            'accessDenied' (for 403) or 'unauthorized' for (401)
        title:
          type: string
          description: detailed description of the error in natural language (english)
    Error422:
      description: >
        The zoneType is only available for certain 422 errors,  to indicate that
        a certain zone specific operation is not allowed because of the ZoneType
        of the zone.
      type: object
      allOf:
        - $ref: "#/components/schemas/Error"
        - properties:
            zoneType:
              $ref: "#/components/schemas/ZoneType"
    ErrorResponse:
      type: object
      description: error object returned for non-200 responses
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    ErrorResponse422:
      type: object
      description: error object returned for 422 responses
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error422"
...
    HeatingCircuitId:
      type: integer
      example: 1
...
    HomeId:
      type: integer
      format: int64
      example: 1118186
      description: unique home ID
...
    Orientation:
      type: string
      enum:
        - HORIZONTAL
        - VERTICAL
      example: HORIZONTAL
      description: tado device property known to be present for DeviceType VA02
...
    Zone:
      type: object
      description: a room in a home whose temperature is controlled by TaNoClo
      properties:
        id:
          $ref: "#/components/schemas/ZoneId"
        name:
          type: string
          description: user assigned name
          example: my living room
        type:
          $ref: "#/components/schemas/ZoneType"
        dateCreated:
          type: string
          format: date-time
          example: 2022-11-25T09:40:14.803Z
        deviceTypes:
          type: array
          items:
            $ref: "#/components/schemas/DeviceType"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceExtra"
        reportAvailable:
          type: boolean
          example: false
        showScheduleSetup:
          type: boolean
          example: true
        supportsDazzle:
          type: boolean
          example: true
        dazzleEnabled:
          type: boolean
          example: true
        dazzleMode:
          type: object
          properties:
            supported:
              type: boolean
              example: true
            enabled:
              type: boolean
              example: true
        openWindowDetection:
          type: object
          properties:
            supported:
              type: boolean
              example: true
            enabled:
              type: boolean
              example: true
            timeoutInSeconds:
              type: integer
              example: 900
...
    ZoneControl:
      description: >
        Result of /homes/{homeId}/zones/{zoneId}/control.


        Groups the devices in a given zone by their duty.

        The elements included in the result depend on the ZoneType:

        * AIR_CONDITIONING: _unknown_

        * HEATING: includes earlyStartEnabled, heatingCircuit, duties.leader,
        duties.drivers and duties.uis

        * HOT_WATER: includes duties.driver and duties.ui
      type: object
      properties:
        type:
          $ref: "#/components/schemas/ZoneType"
        earlyStartEnabled:
          type: boolean
          example: true
          description: applicable for ZoneType HEATING
        heatingCircuit:
          $ref: "#/components/schemas/HeatingCircuitId"
        duties:
          type: object
          properties:
            type:
              $ref: "#/components/schemas/ZoneType"
            driver:
              $ref: "#/components/schemas/Device"
            drivers:
              type: array
              items:
                $ref: "#/components/schemas/Device"
            leader:
              $ref: "#/components/schemas/Device"
            leaders:
              type: array
              items:
                $ref: "#/components/schemas/Device"
            ui:
              $ref: "#/components/schemas/Device"
            uis:
              type: array
              items:
                $ref: "#/components/schemas/Device"
    ZoneDetailsInput:
      description: Used in PUT /homes/{homeId}/zones/{zoneId}/details
      type: object
      properties:
        name:
          type: string
          example: bed room
    ZoneId:
      type: integer
      example: 3
      description: ID of a zone (unique only within the home it belongs to)
...
    ZoneType:
      type: string
      enum:
        - AIR_CONDITIONING
        - HEATING
        - HOT_WATER
      example: HEATING
...